MoonCats: use pre-generated images on IPFS
