+++ package.json
@@ -48,8 +48,5 @@
     "react-dom": "^17.0.1",
     "rollpkg": "^0.5.2",
     "typescript": "^4.2.3"
-  },
-  "dependencies": {
-    "mooncatparser": "^1.0.0"
   }
 }
