+++ src/fetchers/ethers/mooncats.ts
@@ -1,8 +1,7 @@
 import type { Address, NftMetadata } from "../../types"
 import type { EthersFetcherOptions } from "./types"
 
-import mooncat from "mooncatparser"
-import { addressesEqual } from "../../utils"
+import { addressesEqual, fetchImage, ipfsUrl } from "../../utils"
 
 const MOONCATS_WRAPPED_CONTRACT = "0x7c40c393dc0f283f318791d746d894ddd3693572"
 
@@ -10,39 +9,38 @@ const MOONCATS_WRAPPED_ABI = [
   "function _tokenIDToCatID(uint256 tokenId) view returns (bytes5 catId)",
 ]
 
-const CANVAS_SIZE = 10
+// See https://www.reddit.com/r/MoonCatRescue/comments/m5d7mx/svg_imagery_of_all_rescued_mooncats/
+const MOONCATS_IPFS_CID =
+  "bafybeidk4zunuq56w2pf2sncexohlyqae62dzplljkbwswa7jwywh2dava"
 
-function imageUrl(catId: string): string | null {
-  const data = mooncat(catId)
+async function imageUrl(catId: string): Promise<string | null> {
+  const dir = catId.slice(4, 6)
+  const url = ipfsUrl(`ipfs://ipfs/${MOONCATS_IPFS_CID}/${dir}/${catId}.png`)
 
-  const width = CANVAS_SIZE * data.length
-  const height = CANVAS_SIZE * data[0].length
+  const image = await fetchImage(url)
+  const width = image.naturalWidth * 4
+  const height = image.naturalHeight * 4
+  const hPadding = width * 0.125
+  const vPadding = height * 0.125
 
   const canvas = document.createElement("canvas")
   canvas.width = width
   canvas.height = height
 
-  const ctx = canvas.getContext("2d", {
-    antialias: false,
-  }) as CanvasRenderingContext2D
+  const ctx = canvas.getContext("2d")
 
   if (ctx === null) {
     return null
   }
 
-  data.forEach((cells: string[], col: number) => {
-    cells.forEach((color: string, row: number) => {
-      if (color) {
-        ctx.fillStyle = color
-        ctx.fillRect(
-          Math.floor((col * CANVAS_SIZE) / 2 + width / 4),
-          Math.floor((row * CANVAS_SIZE) / 2 + height / 4),
-          Math.floor(CANVAS_SIZE / 2),
-          Math.floor(CANVAS_SIZE / 2)
-        )
-      }
-    })
-  })
+  ctx.imageSmoothingEnabled = false
+  ctx.drawImage(
+    image,
+    hPadding,
+    vPadding,
+    width - hPadding * 2,
+    height - vPadding * 2
+  )
 
   return canvas.toDataURL()
 }
@@ -58,7 +56,7 @@ export async function moonCatsMetadata(
   )
 
   const catId = await wrappedContract._tokenIDToCatID(tokenId)
-  const image = imageUrl(catId) ?? ""
+  const image = (await imageUrl(catId)) ?? ""
 
   return {
     name: `Wrapped MoonCat #${tokenId}`,
